feat: refactor and improive
