--- packages/api/utils/global.ts
@@ -1,3 +1,0 @@
-import * as process from 'process';
-
-export const chatDocumentsFolder = process.env.CHAT_DOCUMENTS_FOLDER;
