+++ packages/web-ui/components/chat-tools.tsx
@@ -73,7 +73,7 @@ export function ChatTools({ documents, roomId, isOwner }: ChatToolsProps) {
   };
 
   return (
-    <div className="flex h-full min-w-[220px] flex-col border-l">
+    <div className="flex h-full min-w-[220px] max-w-[220px] flex-col border-l">
       <ScrollArea className="flex-1">
         <div className="space-y-1">
           <Accordion type="single" collapsible className="w-full">
